# By Jorgensen, Truls
# Via Jorgensen, Truls
* commit '6d14d48e38ce889c3e104c1e1d076a4038049812':
  removed console.log
  avoided startup complaining
  superuser and tests
  added some tests, helped me identify some problems
  AURA-497 Superuser mode med Ace editor, yay
  Superuser take 1

+++ war/src/main/java/no/nav/aura/basta/rest/OrdersRestService.java
@@ -9,13 +9,7 @@ import java.util.Collections;
 
 import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
+import javax.ws.rs.*;
 import javax.ws.rs.core.CacheControl;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
@@ -23,6 +17,7 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.ResponseBuilder;
 import javax.ws.rs.core.UriInfo;
 import javax.xml.bind.JAXBException;
+import javax.xml.bind.UnmarshalException;
 
 import no.nav.aura.basta.User;
 import no.nav.aura.basta.backend.FasitUpdateService;
@@ -61,6 +56,7 @@ import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import org.xml.sax.SAXParseException;
 
 @SuppressWarnings("serial")
 @Component
@@ -114,7 +110,7 @@ public class OrdersRestService {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public OrderDO postOrder(OrderDetailsDO orderDetails, @Context UriInfo uriInfo) {
+    public OrderDO postOrder(OrderDetailsDO orderDetails, @Context UriInfo uriInfo, @QueryParam("prepare") Boolean prepare) {
         checkAccess(orderDetails);
         String currentUser = User.getCurrentUser().getName();
         Order order = orderRepository.save(new Order(orderDetails.getNodeType()));
@@ -123,23 +119,71 @@ public class OrdersRestService {
         Settings settings = new Settings(order, orderDetails);
         OrchestatorRequest request = new OrderV2Factory(settings, currentUser, vmInformationUri, resultUri, fasitRestClient).createOrder();
         WorkflowToken workflowToken;
-        if (request instanceof ProvisionRequest) {
-            workflowToken = orchestratorService.send(request);
-        } else if (request instanceof DecomissionRequest) {
-            workflowToken = orchestratorService.decommission((DecomissionRequest) request);
-            Optional<String> hosts = settings.getProperty(DecommissionProperties.DECOMMISSION_HOSTS_PROPERTY_KEY);
-            if (hosts.isPresent()) {
-                fasitUpdateService.removeFasitEntity(order, hosts.get());
+        if (prepare == null || !prepare){
+            if (request instanceof ProvisionRequest) {
+                workflowToken = orchestratorService.send(request);
+                order.setOrchestratorOrderId(workflowToken.getId());
+            } else if (request instanceof DecomissionRequest) {
+                workflowToken = orchestratorService.decommission((DecomissionRequest) request);
+                order.setOrchestratorOrderId(workflowToken.getId());
+                Optional<String> hosts = settings.getProperty(DecommissionProperties.DECOMMISSION_HOSTS_PROPERTY_KEY);
+                if (hosts.isPresent()) {
+                    fasitUpdateService.removeFasitEntity(order, hosts.get());
+                }
+            } else {
+                throw new RuntimeException("Unknown request type " + request.getClass());
             }
-        } else {
-            throw new RuntimeException("Unknown request type " + request.getClass());
+            order.setRequestXml(convertXmlToString(censore(request)));
+        }else{
+            order.setRequestXml(convertXmlToString(request));
         }
-        order.setRequestXml(convertXmlToString(censore(request)));
-        order.setOrchestratorOrderId(workflowToken.getId());
         order = orderRepository.save(order);
         settingsRepository.save(settings);
         return new OrderDO(order, uriInfo);
     }
+
+
+    @PUT
+    @Consumes(MediaType.TEXT_PLAIN)
+    @Produces(MediaType.APPLICATION_JSON)
+    @Path("{orderId}")
+    public Response putXMLOrder(String string, @PathParam("orderId") Long orderId, @Context UriInfo uriInfo) {
+        ProvisionRequest request;
+        try{
+        request = XmlUtils.parseAndValidateXmlString(ProvisionRequest.class, string);
+        }catch(UnmarshalException e){
+            SAXParseException spe = (SAXParseException) e.getLinkedException();
+           return Response.status(400).entity(getValidationMessage(spe)).header("Content-type", "text/plain").build();
+        }
+
+        checksuperDuperAccess(User.getCurrentUser());
+        WorkflowToken workflowToken;
+        //if (request instanceof ProvisionRequest) {
+            workflowToken = orchestratorService.send(request);
+       /* } else {
+            throw new RuntimeException("Unknown request type " + request.getClass());
+        }*/
+
+        Order order = orderRepository.findOne(orderId);
+        if (order.getOrchestratorOrderId() == null){
+            order.setRequestXml(convertXmlToString(censore(request)));
+            order.setOrchestratorOrderId(workflowToken.getId());
+            order = orderRepository.save(order);
+            Settings settings = settingsRepository.findByOrderId(orderId);
+            settings.setXmlCustomized();
+            settingsRepository.save(settings);
+        }
+        return Response.ok(new OrderDO(order, uriInfo)).build();
+    }
+
+    private String getValidationMessage(SAXParseException spe) {
+        String msg = spe.getLocalizedMessage();
+        if (msg.contains(":")){
+           msg = msg.substring(msg.indexOf(":") + 1);
+        }
+        return "(linje "+ spe.getLineNumber() + ", kolonne " + spe.getColumnNumber() + ") " +  msg;
+    }
+
 
     /**
      * @param request
@@ -217,7 +261,11 @@ public class OrdersRestService {
     @Path("{orderId}/requestXml")
     @Produces(MediaType.TEXT_XML)
     public String getRequestXml(@PathParam("orderId") long orderId) {
-        return orderRepository.findOne(orderId).getRequestXml();
+        Order order = orderRepository.findOne(orderId);
+        if (order.getOrchestratorOrderId() == null){
+            checksuperDuperAccess(User.getCurrentUser());
+        }
+        return order.getRequestXml();
     }
 
     @GET
@@ -274,5 +322,12 @@ public class OrdersRestService {
         // TODO Check remote address
         logger.info("Called from " + remoteAddr);
     }
+
+    private void checksuperDuperAccess(User currentUser) {
+        if (!currentUser.hasSuperUserAccess()){
+            throw new UnauthorizedException("User " + User.getCurrentUser().getName() + " does not have super user access");
+        }
+    }
+
 
 }
